criando endpoints de /tipos-transportes e cadastro de meio transportes

+++ src/main/java/br/com/aps/simulador/ambiental/repository/TipoTransporteRepository.java
@@ -3,5 +3,5 @@ package br.com.aps.simulador.ambiental.repository;
 import br.com.aps.simulador.ambiental.model.TipoTransporte;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoTransporteRepository extends JpaRepository<TipoTransporte, Integer> {
+public interface TipoTransporteRepository extends JpaRepository<TipoTransporte, Long> {
 }
